// apps/main/src/models/v2/iam/serviceUserSummary.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 */
import type { ServiceUserRole } from './serviceUserRole';

/**
 * Summary fields of the service user
 */
export interface ServiceUserSummary {
  /** The service token id */
  id: string;
  /** The time the service token was created or last regenerated */
  generated_at: string;
  /** The user that created or last regenerated the service token */
  generated_by: string;
  /** A display name of the service user */
  display_name: string;
  /** The time at which the current generation of the token expires */
  current_token_expires_at: string;
  /** The list of roles assigned to this service user */
  roles: ServiceUserRole[];
}